+++ tests/variants.test.js
@@ -134,6 +134,39 @@ describe('custom advanced variants', () => {
       `)
     })
   })
+
+  test('using variants with multi-class selectors', () => {
+    let config = {
+      content: [
+        {
+          raw: html` <div class="screen:parent screen:child"></div> `,
+        },
+      ],
+      plugins: [
+        function ({ addVariant, addComponents }) {
+          addComponents({
+            '.parent .child': {
+              foo: 'bar',
+            },
+          })
+          addVariant('screen', '@media screen')
+        },
+      ],
+    }
+
+    return run('@tailwind components;@tailwind utilities', config).then((result) => {
+      return expect(result.css).toMatchFormattedCss(css`
+        @media screen {
+          .screen\:parent .child {
+            foo: bar;
+          }
+          .parent .screen\:child {
+            foo: bar;
+          }
+        }
+      `)
+    })
+  })
 })
 
 test('stacked peer variants', async () => {
@@ -286,6 +319,66 @@ test('custom addVariant with nested media & format shorthand', () => {
             text-align: center;
           }
         }
+      }
+    `)
+  })
+})
+
+test('before and after variants are a bit special, and forced to the end', () => {
+  let config = {
+    content: [
+      {
+        raw: html`
+          <div class="before:hover:text-center"></div>
+          <div class="hover:before:text-center"></div>
+        `,
+      },
+    ],
+    plugins: [],
+  }
+
+  return run('@tailwind components;@tailwind utilities', config).then((result) => {
+    return expect(result.css).toMatchFormattedCss(css`
+      .before\:hover\:text-center:hover::before {
+        content: var(--tw-content);
+        text-align: center;
+      }
+
+      .hover\:before\:text-center:hover::before {
+        content: var(--tw-content);
+        text-align: center;
+      }
+    `)
+  })
+})
+
+test('before and after variants are a bit special, and forced to the end (2)', () => {
+  let config = {
+    content: [
+      {
+        raw: html`
+          <div class="before:prose-headings:text-center"></div>
+          <div class="prose-headings:before:text-center"></div>
+        `,
+      },
+    ],
+    plugins: [
+      function ({ addVariant }) {
+        addVariant('prose-headings', ':where(&) :is(h1, h2, h3, h4)')
+      },
+    ],
+  }
+
+  return run('@tailwind components;@tailwind utilities', config).then((result) => {
+    return expect(result.css).toMatchFormattedCss(css`
+      :where(.before\:prose-headings\:text-center) :is(h1, h2, h3, h4)::before {
+        content: var(--tw-content);
+        text-align: center;
+      }
+
+      :where(.prose-headings\:before\:text-center) :is(h1, h2, h3, h4)::before {
+        content: var(--tw-content);
+        text-align: center;
       }
     `)
   })